fix(vite): dev ssr fouc
