+++ packages/vite/src/client.ts
@@ -1,6 +1,5 @@
 import { resolve } from 'pathe'
 import * as vite from 'vite'
-import fse from 'fs-extra'
 import consola from 'consola'
 import vitePlugin from '@vitejs/plugin-vue'
 
@@ -8,6 +7,7 @@ import { cacheDirPlugin } from './plugins/cache-dir'
 import { replace } from './plugins/replace'
 import { wpfs } from './utils/wpfs'
 import type { ViteBuildContext, ViteOptions } from './vite'
+import { writeManifest } from './manifest'
 
 export async function buildClient (ctx: ViteBuildContext) {
   const clientConfig: vite.InlineConfig = vite.mergeConfig(ctx.config, {
@@ -68,24 +68,5 @@ export async function buildClient (ctx: ViteBuildContext) {
     consola.info(`Client built in ${Date.now() - start}ms`)
   }
 
-  // Write client manifest for use in vue-bundle-renderer
-  const clientDist = resolve(ctx.nuxt.options.buildDir, 'dist/client')
-  const serverDist = resolve(ctx.nuxt.options.buildDir, 'dist/server')
-
-  // Legacy dev manifest
-  const devClientManifest = {
-    publicPath: ctx.nuxt.options.build.publicPath,
-    all: ['@vite/client', 'entry.mjs'],
-    initial: ['@vite/client', 'entry.mjs'],
-    async: [],
-    modules: {}
-  }
-
-  const clientManifest = ctx.nuxt.options.dev
-    ? devClientManifest
-    : await fse.readJSON(resolve(clientDist, 'manifest.json'))
-
-  await fse.mkdirp(serverDist)
-  await fse.writeFile(resolve(serverDist, 'client.manifest.json'), JSON.stringify(clientManifest, null, 2), 'utf8')
-  await fse.writeFile(resolve(serverDist, 'client.manifest.mjs'), 'export default ' + JSON.stringify(clientManifest, null, 2), 'utf8')
+  await writeManifest(ctx)
 }
